Replace python-plexapi with custom Json service
Tidy ups, fixes and improvements
