+++ mediux_posters/services/plex/schemas.py
@@ -1,14 +1,10 @@
-__all__ = ["Collection", "Episode", "Movie", "Season", "Show"]
+__all__ = ["Collection", "Episode", "Library", "Movie", "Season", "Show"]
 
+from datetime import date
+from enum import Enum
 
-from plexapi.collection import Collection as PlexCollection
-from plexapi.video import (
-    Episode as PlexEpisode,
-    Movie as PlexMovie,
-    Season as PlexSeason,
-    Show as PlexShow,
-)
-from pydantic import Field
+from pydantic import Field, model_validator
+from pydantic.alias_generators import to_camel
 
 from mediux_posters.services._base import (
     BaseCollection,
@@ -17,24 +13,108 @@ from mediux_posters.services._base import (
     BaseSeason,
     BaseShow,
 )
+from mediux_posters.utils import BaseModel
 
 
-class Episode(BaseEpisode, arbitrary_types_allowed=True):
-    plex: PlexEpisode | None = None
+class PlexModel(BaseModel, alias_generator=to_camel, extra="ignore"): ...
 
 
-class Season(BaseSeason, arbitrary_types_allowed=True):
-    plex: PlexSeason | None = None
+class MediaType(str, Enum):
+    ARTIST = "artist"
+    MOVIE = "movie"
+    PHOTO = "photo"
+    SHOW = "show"
 
 
-class Show(BaseShow, arbitrary_types_allowed=True):
-    plex: PlexShow | None = None
+class Library(PlexModel):
+    id: int = Field(alias="key")
+    title: str
+    type: MediaType
 
 
-class Movie(BaseMovie, arbitrary_types_allowed=True):
-    plex: PlexMovie | None = None
+class Episode(BaseEpisode, PlexModel):
+    id: int = Field(alias="ratingKey")
+    name: str = Field(alias="title")
+    number: int = Field(alias="index")
+    premiere_date: date | None = Field(alias="originallyAvailableAt", default=None)
+
+    @model_validator(mode="before")
+    def extract_guids(cls, data: object) -> object:
+        if isinstance(data, dict) and "Guid" in data:
+            services = {
+                f"{x['id'].split('://')[0]}_id": x["id"].split("://")[1]
+                for x in data["Guid"]
+                if "id" in x
+            }
+            data.update(services)
+            del data["Guid"]
+        return data
 
 
-class Collection(BaseCollection, arbitrary_types_allowed=True):
-    movies: list[Movie] = Field(default_factory=list)
-    plex: PlexCollection | None = None
+class Season(BaseSeason, PlexModel):
+    id: int = Field(alias="ratingKey")
+    name: str = Field(alias="title")
+    number: int = Field(alias="index")
+
+    @model_validator(mode="before")
+    def extract_guids(cls, data: object) -> object:
+        if isinstance(data, dict) and "Guid" in data:
+            services = {
+                f"{x['id'].split('://')[0]}_id": x["id"].split("://")[1]
+                for x in data["Guid"]
+                if "id" in x
+            }
+            data.update(services)
+            del data["Guid"]
+        return data
+
+
+class Show(BaseShow, PlexModel):
+    id: int = Field(alias="ratingKey")
+    name: str = Field(alias="title")
+    premiere_date: date = Field(alias="originallyAvailableAt")
+
+    @model_validator(mode="before")
+    def extract_guids(cls, data: object) -> object:
+        if isinstance(data, dict) and "Guid" in data:
+            services = {
+                f"{x['id'].split('://')[0]}_id": x["id"].split("://")[1]
+                for x in data["Guid"]
+                if "id" in x
+            }
+            data.update(services)
+            del data["Guid"]
+        return data
+
+
+class Movie(BaseMovie, PlexModel):
+    id: int = Field(alias="ratingKey")
+    name: str = Field(alias="title")
+    premiere_date: date = Field(alias="originallyAvailableAt")
+
+    @model_validator(mode="before")
+    def extract_guids(cls, data: object) -> object:
+        if isinstance(data, dict) and "Guid" in data:
+            services = {
+                f"{x['id'].split('://')[0]}_id": x["id"].split("://")[1]
+                for x in data["Guid"]
+                if "id" in x
+            }
+            data.update(services)
+            del data["Guid"]
+        return data
+
+
+class Collection(BaseCollection, PlexModel):
+    id: int = Field(alias="ratingKey")
+    name: str = Field(alias="title")
+
+    @model_validator(mode="before")
+    def extract_guids(cls, data: object) -> object:
+        if isinstance(data, dict) and "Label" in data:
+            if tag := next(iter(x for x in data["Label"] if x["tag"].startswith("Tmdb-")), {}).get(
+                "tag"
+            ):
+                data["tmdb_id"] = int(tag.removeprefix("Tmdb-"))
+            del data["Label"]
+        return data
